requirements.txt
flask
torch
torchvision
onnxruntime
numpy
opencv-python
Pillow
